chore: move folders into shared
