--- src/procedural_algorithm.rs
@@ -1,2 +1,0 @@
-pub mod perlin_noise_cave;
-pub mod random_walk_cave;
