+++ src/shared/procedural_map/rectangle_spawning.rs
