+++ src/shared/procedural_map/rectangle_spawning/rectangle_entity.rs
